+++ packages/@webex/plugin-meetings/src/meetings/meetings.types.ts
@@ -3,7 +3,10 @@ import type {
   VirtualBackgroundEffectOptions,
 } from '@webex/media-helpers';
 
-type INoiseReductionEffect = Omit<NoiseReductionEffectOptions, 'authToken'>;
+type INoiseReductionEffect = Omit<
+  NoiseReductionEffectOptions,
+  'authToken' | 'workletProcessorUrl' | 'legacyProcessorUrl'
+>;
 type IVirtualBackgroundEffect = Omit<VirtualBackgroundEffectOptions, 'authToken'>;
 
 export type {INoiseReductionEffect, IVirtualBackgroundEffect};